refactore - Reorganaze a folder structure to add a layouts folder
