+++ src/sections/promo/Promo.jsx
@@ -61,7 +61,6 @@ const Promo = () => {
                 width="500"
                 height="500"
                 alt={'A funny pixel men with a laptop in his hands'}
-                // loading="eager"
                 fetchPriority='high'
                 decoding='async'
               />
